Very basic rendering
Added button to MainWindow that opens the MeshView with a scene that
just contains a triangle. See MainWindow::openMeshView().

Added code to main.cpp to set the QSurfaceFormat default format. This
seems to be necessary for rendering OpenGL (at least on some systems
(?)).

Camera can be dragged in MeshView (in XY plane) using the mouse.

The rendered mesh is always purple.

Not yet tested with anything more complicated than a triangle.