Tool & ToolManager, also DrawableGLObject with a DrawableAxes example (#6)
Click and drag with the left mouse button to rotate view. There is no
way to pan just yet.

* Camera panning

Added camera panning. Hold shift when you start dragging to pan the
camera.

* base on tima's branch, added obj loading function, a bug fixed.

* remove .pro.user file

* fresh to death with onion breath

* Fix panning

Fixed a miscalculation in the panning code that made it misbehave a
little.

* DrawableGLObject and axes

Added the DrawableGLObject. This is any object that needs special
rendering. They can be added to a Scene just like RenderableObjects.

Note for future: I have noticed occasional crashes happening right when
the Show View button is pressed in the UI. These happen infrequently
and without a pattern.

* Removed constructor from DrawableGLObject

It is an abstract class that should not be default-constructible.

* Tool and ToolManager

Tool is an abstract class representing a UI tool (like a brush).

ToolManager represents a group of mutually exclusive tools, e.g.
“rectangle”, “line”, “erase”.

* Added files to WallsAndHoles.pro (messed up in merge)

* base on tima's branch, added obj loading function, a bug fixed.

* remove .pro.user file

* Fix panning

Fixed a miscalculation in the panning code that made it misbehave a
little.

* DrawableGLObject and axes

Added the DrawableGLObject. This is any object that needs special
rendering. They can be added to a Scene just like RenderableObjects.

Note for future: I have noticed occasional crashes happening right when
the Show View button is pressed in the UI. These happen infrequently
and without a pattern.

* Removed constructor from DrawableGLObject

It is an abstract class that should not be default-constructible.

* Tool and ToolManager

Tool is an abstract class representing a UI tool (like a brush).

ToolManager represents a group of mutually exclusive tools, e.g.
“rectangle”, “line”, “erase”.

* Added Abstract prefix to abstract classes

Tool -> AbstractTool
MeshViewCamera -> AbstractMeshViewCamera
DrawableGLObject -> AbstractDrawableGLObject

* Amend to previous commit

Forgot to include AbstractDrawableGLObject

* Better panning

Camera panning now considers zoom level.

* Example Tool Usage, MeshView Container Window

MeshViewContainer class represents the window that contains the
MeshView. This is so that UI elements (such as a tool selection) can be
added.

The Blender-like camera now inherits from AbstractTool. See
MeshView::MeshView() for example code using the camera along with a
tool manager. Also see MeshViewContainer.

* ToolManager returns QAction

ToolManager::registerTool() returns a QAction. I have not tested this
yet, but it should work.