Improved Mesh Generation
M2MTileMesher now outputs boxes, and takes in an M2MPropertySet.

Map2Mesh has an inferProperties() method. For now, this just computes
the “base” and “top” height of each tile (see below).

Map2Mesh::Properties will be used to predefine types of properties.
Currently, there is only Map2Mesh::Properties::Height, which keeps
track of the “top height” and “base height” of a tile (the base height
is the minimum height of the neighbors of a tile).

Added TileMap::getArray2D() convenience method.