Changed SimpleTexturedRenderer to use SimpleTexturedShader
Also added a mutex for modifying or using data.

Rendering order is now determined by texture to minimize the number of
texture calls.

Added extra functions for dealing with textures.

SimpleTexturedObject now uses QImages instead of QOpenGLTextures. This
is because QOpenGLTextures need to be created in an OpenGL context, so
the natural place to create them is inside a renderer’s initializeGL().

M2MTileMesher now outputs a textured mesh.

Added cleanUp() to MeshView to free resources when context is changed.