Enabled back-face culling.
Note: Map2Mesh still outputs faces based on a “base” and “top” height,
where the base is the minimum of all the surrounding top heights.
Therefore, if a tile has at least one neighboring tile with a lower
height, it will draw all of its vertical faces with that height.