Made M2M namespace, and remade TileMesher system
At the time of this commit, the program should compile and work
correctly, using the updated meshing system.

Instead of an “M2MTileMesher” class, there is now an M2M namespace with
a TileMesher class.

A TileMesher does two things:

1) Keeps information about the tile neighborhood it was created with
for the sake of being able to check whether the tile’s mesh may need an
update.

2) Has a makeMesh() method that returns a list of meshes that represent
the tile. This is an abstract method implemented by only one subclass
currently, but I will be adding a second subclass (TileDiagonalMesher)
shortly.