Diagonals and bridging (#69)
Main highlight: Diagonals and Bridging between tiles of the same template set is now supported!
This is a template level feature, and is changeable from the Template Set property browser. Bridging and diagonals can be set exclusively of each other.

General outline of changes:

Some systemic changes to the Map2Mesh system
Tiles all use one of two tileMeshers:
BlockyPolygonTileMesher - Meshes a non-ground tile
GroundBlockyPolygonTileMesher - Meshes a ground tile
Each mesh everything needed within the bounds of the tile's grid.
Both meshers are based on an AbstractPolygonTileMesher
Notes:

May not be perfect. A lot was rushed, and involves a lot of cases, and I likely did not test them all.
Though there is code to support it, changing a tiles position caused a lot of issues with diagonals and bridging, and some crashing. There are also other areas of the program that don't handle position too well. Based on this, and the fairly minimal gains of allowing positions, they have been disabled (just from the UI though).
There can be some perhaps undesired (but not buggy) behavior when using diagonals. Mainly revolving around there being diagonals where you might not want them. Not an actual issue, just noting we will eventually want the ability to on a case by case basis choose whether or not a diagonal forms.